Improvements to DAO classes

See merge request markussp/idatt1002_2021_28!46

+++ src/main/java/ntnu/idatt1002/dao/UserDAO.java
@@ -10,14 +10,14 @@ import java.security.spec.KeySpec;
 import java.util.ArrayList;
 
 /**
- * Access user objects in storage
+ * Static class to access user objects in storage
  */
 public final class UserDAO {
     private static final String SAVEPATH = "src/main/resources/saves";
     private static final String FILETYPE = ".ser";
 
     /**
-     * Get all users in storage
+     * Get a list of all users in storage
      * @return an {@code ArrayList} of {@code User} objects
      */
     public static ArrayList<User> getUsers(){
@@ -28,6 +28,7 @@ public final class UserDAO {
         FilenameFilter filter = (directory1, name) -> !name.endsWith(".ser") && !name.contains(".gitkeep");
         String[] pathnames = saveDirectory.list(filter);
 
+        //Remaining files should be user folders
         if(pathnames != null){
             for(String path : pathnames){
                 User user = deserializeUser(path);
@@ -39,7 +40,7 @@ public final class UserDAO {
     }
 
     /**
-     * Save user to storage
+     * Save user to storage. Will overwrite if equal user already is stored.
      * @param user {@code User} object
      */
     public static void serializeUser(User user){
@@ -57,10 +58,13 @@ public final class UserDAO {
             if(!success){ System.out.println("Error occured"); }
         }
 
+        //Write to file
         try {
             FileOutputStream fos = new FileOutputStream(file);
             ObjectOutputStream oos = new ObjectOutputStream(fos);
+
             oos.writeObject(user);
+
             oos.close();
             fos.close();
         } catch (IOException ioe) {
@@ -70,12 +74,17 @@ public final class UserDAO {
     }
 
     /**
-     * Get user object from storage
-     * @return {@code null} if error occurs
+     * Get user object from storage.
+     * @param username non case-sensitive username
+     * @return User object with given username, {@code null} if user could not be found.
      */
     public static User deserializeUser(String username){
         File file = new File(filePath(username));
         User user = null;
+
+        if(!file.exists()){
+            return null;
+        }
         try{
             FileInputStream fis = new FileInputStream(file);
             ObjectInputStream ois = new ObjectInputStream(fis);
@@ -91,10 +100,13 @@ public final class UserDAO {
     }
 
     /**
-     * Delete a user and all its files
+     * Delete a user and all its files.
+     * @param username non case-sensitive username
      * @return {@code false} if the user folder or some of its elements could not be deleted
      */
     public static boolean deleteUser(String username){
+        if(!userExists(username)){ return false; }
+
         //Tasks & categories
         boolean success = CategoryDAO.deleteCategoriesByUser(username);
         File categoryDir = new File(userDir(username) + "Categories");
@@ -113,9 +125,20 @@ public final class UserDAO {
         return success;
     }
 
+    /**
+     * Check if given user exists in storage
+     * @param username non case-sensitive username
+     * @return true or false
+     */
+    static boolean userExists(String username){
+        File userDir = new File(userDir(username));
+        return userDir.exists();
+    }
+
     //PASSWORD SALT & HASHING
     /**
-     * @return a random salt
+     * Generate a random salt to store in a user.
+     * @return a random salt.
      */
     public static byte[] generateSalt() {
         SecureRandom random = new SecureRandom();
@@ -125,10 +148,11 @@ public final class UserDAO {
     }
 
     /**
-     * PBKDF2 method to hash a password with salt
+     * PBKDF2 method to hash a password with salt. Needs equal salt and equal password to produce
+     * the same has.
      * @param password password in clear-text
      * @param salt a users personal salt
-     * @return {@code null} if an error occured
+     * @return Hashed password in String format, {@code null} if an error occured.
      */
     public static String hashPassword(String password, byte[] salt){
         KeySpec spec = new PBEKeySpec(password.toCharArray(), salt, 65536, 128);
